refactor: move actions and listeners to separated files
